+++ src/extension.ts
@@ -828,7 +828,7 @@ export function activate(context: vscode.ExtensionContext) {
                 URIs.push(vscode.window.activeTextEditor?.document.uri);
             }
             if (URIs.length === 0) {
-                Extension.showErrorMessage("Can't find files for downloading");
+                Extension.showErrorMessage("Can't find files for deleting");
                 return;
             }
             if (Targets.getActive().length === 0) {
@@ -846,6 +846,75 @@ export function activate(context: vscode.ExtensionContext) {
                         } else {
                             target.deleteDir(uri);
                         }
+                    });
+                });
+            });
+        })
+    );
+    context.subscriptions.push(
+        vscode.commands.registerCommand("pro-deployer.delete-from", (...args) => {
+            const URIs = [] as vscode.Uri[];
+            args.forEach((arg) => {
+                if (arg instanceof vscode.Uri) {
+                    const checkExist = URIs.find((item) => {
+                        return item.toString() === arg.toString();
+                    });
+                    if (!checkExist) {
+                        URIs.push(arg);
+                    }
+                }
+                if ("resourceUri" in arg) {
+                    const checkExist = URIs.find((item) => {
+                        return item.toString() === arg.resourceUri.toString();
+                    });
+                    if (!checkExist) {
+                        URIs.push(arg.resourceUri);
+                    }
+                }
+                if (Array.isArray(arg)) {
+                    arg.forEach((uri) => {
+                        if (uri instanceof vscode.Uri) {
+                            const checkExist = URIs.find((item) => {
+                                return item.toString() === uri.toString();
+                            });
+                            if (!checkExist) {
+                                URIs.push(uri);
+                            }
+                        }
+                    });
+                }
+            });
+            if (URIs.length === 0 && vscode.window.activeTextEditor?.document.uri) {
+                URIs.push(vscode.window.activeTextEditor?.document.uri);
+            }
+            if (URIs.length === 0) {
+                Extension.showErrorMessage("Can't find files for deleting");
+                return;
+            }
+
+            const items = Targets.getItems().map((item) => {
+                return {
+                    label: item.getName(),
+                    description: item.getWorkspaceFolder().name,
+                    target: item,
+                };
+            });
+            vscode.window.showQuickPick(items).then((value) => {
+                if (!value) {
+                    return;
+                }
+
+                const target = value.target;
+
+                target.connect(() => {
+                    URIs.forEach((uri) => {
+                        Extension.isLikeFile(uri).then((isFile) => {
+                            if (isFile) {
+                                target.delete(uri);
+                            } else {
+                                target.deleteDir(uri);
+                            }
+                        });
                     });
                 });
             });